--- Veldrid/VkFormats.cs
@@ -1,34 +1,0 @@
-﻿using Vortice.Vulkan;
-
-namespace Veldrid
-{
-    internal static partial class VkFormats
-    {
-        internal static uint GetSpecializationConstantSize(ShaderConstantType type)
-        {
-            switch (type)
-            {
-                case ShaderConstantType.Bool:
-                    return 4;
-                case ShaderConstantType.UInt16:
-                    return 2;
-                case ShaderConstantType.Int16:
-                    return 2;
-                case ShaderConstantType.UInt32:
-                    return 4;
-                case ShaderConstantType.Int32:
-                    return 4;
-                case ShaderConstantType.UInt64:
-                    return 8;
-                case ShaderConstantType.Int64:
-                    return 8;
-                case ShaderConstantType.Float:
-                    return 4;
-                case ShaderConstantType.Double:
-                    return 8;
-                default:
-                    throw Illegal.Value<ShaderConstantType>();
-            }
-        }
-    }
-}
